+++ resources/views/pages/trackings.blade.php
@@ -174,8 +174,8 @@
                                                 @endif
                                             </td>
                                         @endif
-                                        <td class="border-0">{{ $tracking->started_at->format('d/m/Y') }} <strong>{{ $tracking->started_at->format('H:i') }}</strong></td>
-                                        <td class="border-0">{{ $tracking->ended_at->format('d/m/Y') }} <strong>{{ $tracking->ended_at->format('H:i') }}</strong></td>
+                                        <td class="border-0">{{ tz($tracking->started_at)->format('d/m/Y') }} <strong>{{ tz($tracking->started_at)->format('H:i') }}</strong></td>
+                                        <td class="border-0">{{ tz($tracking->ended_at)->format('d/m/Y') }} <strong>{{ tz($tracking->ended_at)->format('H:i') }}</strong></td>
                                         <td class="border-0" data-bs-toggle="tooltip" data-bs-title="{{ $tracking->duration_decimal }}">{{ $tracking->duration }}</td>
                                         @php $bhSeconds = $tracking->duration_seconds - $tracking->non_billable_seconds; @endphp
                                         <td class="border-0" @if($tracking->billable_hours !== null) data-bs-toggle="tooltip" data-bs-title="{{ number_format($bhSeconds / 3600, 2) }}h" @endif>